+++ packages/core/src/core/EventBridge/index.ts
@@ -1,4 +1,5 @@
 export * from './EventBridgeBaseClass.impl.js'
 export * from './InFlightExecutionTracker.impl.js'
 export * from './PendingInvocationRegistry.impl.js'
+export * from './PendingStreamRegistry.impl.js'
 export * from './types/index.js'